+++ src/labelable/__init__.py
@@ -1,3 +1,3 @@
 """Labelable - A general purpose label printing API and UI."""
 
-__version__ = "0.2.1.dev4"
+__version__ = "0.2.1.dev5"